Updates dependencies and supported vscode (#135)
This updates dependencies via Dependabot updates. However, in doing so,
this also had to update the `vscode` engine, which means that the plugin
is now compatible with 1.72.0 and all minor and patch versions equal or
greater. This was made necessary by the upgrade to `@types/vscode`.

If this is an issue, we can downgrade the types package or ignore these
updates altogether.

This also adds a few minor cleanup fixes:
- `vscode-test` was renamed to `@vscode/test-electron`
- Adds an `npx` command to execute the node modules binaries.

Verified changes by installing to local installation of VSCode.