Use Buf CLI as an LSP (#392)
Hi team, I've spent the last couple weekends looking at #344 which
appears abandoned. This snowballed a bit into a rewrite for mostly
maintainability and testability reasons.

Apologies for the size of the pull - happy to split it up if you want to
take it on, else I'll continue happily running this locally.

Changes I've introduced:
  - Introduce a command pattern for command handling.
- Rewrite installation processes via install/update commands (now works
cross platform).
  - Provide flexible method of using buf located in path/storage.
  - Improve status bar with a command picker.
  - Support 'buf generate' as a command.
  - ~~Support module detection for the active file in editor.~~
  - ~~Provide a shortcut to opening 'buf.yaml' as a command.~~

---------

Co-authored-by: Paul Sachs <psachs@buf.build>