Temporarily disable playwright tests when buf is not on the system $PATH
We are already testing the code path for installation in the integration
tests, where we are able to set up a mock server to grab a buf CLI
binary installed to the workspace via `npm install`.

In the case of playwright tests, everything runs in a sandbox, and we do
not have access to the buf binary set up outside of the test. We'll
temporarily disable these tests until we can find a good solution for
this.